Making elasticsearch document type configurable
Document type was default to the kafka message topic. This change
will make it configurable from the configuration file. If it is
not defined in the configuration file, then the document type
which messages go into will fall back to the topic. For example:

[es_persister]
topic = metrics

In the above configuration, all the messages saved by this
persister will be saved under document type "metrics".

[es_persister]
topic = metrics
doc_type = maples

In the above configuration, all the messages handled by this
persister will be saved under document type "maples".