added coverage support for the project
the project is missing the unit test case coverage. This patchset
will allow developer to run a command to see the unit test case
coverage. To do that, run the following command:
    tox -e cover

If the command finishes successfully, a new directory named cover
at the project root directory will be created. Use your browser
to open up the index.html file to see the coverage.